Fix infrequent walking & cycling routing bug
Cron takes an intersection of day of the week and days of the month so:
1-7 * 2 was running every Sat AND the first 7 days, i.e. a union not an intersection.
Cron does run an intersection if the day of week or day of month starts with *
so prefixing */8 to day of the week (i.e. the 8th day of the week) causes the line to
run on the first 7 days of the month only if they are Sat.
Fixes #16